% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/positions.R
\name{cluster_shell}
\alias{cluster_shell}
\title{cluster_shell}
\usage{
cluster_shell(N = 50, R0 = 30, d = 1, a = 1, b = 1, c = 1,
  position = c("quasi-random", "random"), orientation = c("radial", "flat",
  "random"), seed = 123, ...)
}
\arguments{
\item{N}{number of particles}

\item{R0}{radius of core}

\item{d}{distance}

\item{a}{ellipsoid semi-axis}

\item{b}{ellipsoid semi-axis}

\item{c}{ellipsoid semi-axis}

\item{position}{type of random coverage}

\item{seed}{random seed for reproducibility}

\item{...}{extra arguments (ignored)}

\item{position}{type of angular orientation}
}
\value{
list
}
\description{
cluster_shell
}
\details{
shell cluster
}
\author{
baptiste Auguie
}
\seealso{
Other user_level cluster: \code{\link{cluster_chain}};
  \code{\link{cluster_dimer_end}};
  \code{\link{cluster_dimer}}; \code{\link{cluster_helix}};
  \code{\link{equal_angles}}; \code{\link{equal_sizes}};
  \code{\link{helix}}
}

