% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dispersion.r
\name{dispersion_spectrum}
\alias{dispersion_spectrum}
\title{dispersion_spectrum}
\usage{
dispersion_spectrum(cluster, material, medium = 1.33, angles = 0,
  axes = "z", polarisation = c("linear", "circular"), Niter = 10,
  tol = 1e-05, progress = FALSE)
}
\arguments{
\item{cluster}{list describing a cluster}

\item{material}{list}

\item{medium}{medium refractive index}

\item{angles}{of incident field in radians}

\item{axes}{of incident field rotation character vector from ('x', 'y', 'z')}

\item{polarisation}{linear or circular polarisation}

\item{Niter}{integer maximum order of scattering}

\item{tol}{double, tolerance of solution}

\item{progress}{logical, display progress bar}
}
\value{
data.frame
}
\description{
dispersion spectrum
}
\details{
dispersion spectrum
}
\note{
The incident wavevector is along the z direction.
}
\author{
baptiste Auguie
}

